Salamtuk/requirements.txt
Babel==2.7.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-countries==5.4
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-uuidfield==0.5.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
fcm-django==0.2.21
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.3.3
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
msgpack==0.6.1
oauthlib==3.1.0
phonenumbers==8.10.17
Pillow==6.1.0
protobuf==3.9.1
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyfcm==1.4.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
